Divided Data.js into 2 files
Moved Data.js from main directory to contexts and divided into PokeApiData.js and PokeJSON.js to separate building JSON object from sending requests to API. Also added constant urls.
