+++ src/contexts/PokeJSON.js
@@ -1,0 +1,20 @@
+// Filtering specific pokemon's JSON response and creating custom object:
+export const preparePokemonObject = (jsons) => {
+  // Pokemon hasn't been found:
+  if (jsons.includes('Not found')) return;
+
+  // Pokemon has been found:
+  return {
+    id: jsons[0].id,
+    name: jsons[0].name.charAt(0).toUpperCase() + jsons[0].name.slice(1),
+    height: Number(jsons[0].height) / 10,
+    weight: Number(jsons[0].weight) / 10,
+    sprite: jsons[0].sprites.other['official-artwork'].front_default,
+    types: jsons[0].types.map((i) => i.type).map((i) => i.name),
+    stats: jsons[0].stats.map((i) => [i.stat.name, i.base_stat]),
+    //moves: jsons[0].moves, // This one needs a lot of reshaping, maybe it should be in another function?
+    color: jsons[1].color.name,
+    generation: jsons[1].generation.name.split('-')[1].toUpperCase(),
+    evolution_chain: jsons[1].evolution_chain.url,
+  };
+};
